+++ fluidSimLib/pointGenerator3.h
@@ -19,13 +19,13 @@ public:
 
     void generate(
         const boundingBox3& boundingBox,
-        FloatType spacing,
+        double spacing,
         std::vector<vector3>* points) const;
 
 
     virtual void forEachPoint(
         const boundingBox3& boundingBox,
-        FloatType spacing,
+        double spacing,
         const std::function<bool(const vector3&)>& callback) const = 0;
 };
