+++ fluidSimLib/quaternion.cpp
@@ -8,7 +8,7 @@ quaternion::quaternion()
     setIdentity();
 }
 
-quaternion::quaternion(FloatType newW, FloatType newX, FloatType newY, FloatType newZ)
+quaternion::quaternion(double newW, double newX, double newY, double newZ)
 {
     set(newW, newX, newY, newZ);
 }
@@ -18,21 +18,21 @@ quaternion::quaternion(const matrix33& m33)
     set(m33);
 }
 
-void quaternion::set(FloatType newW, FloatType newX, FloatType newY, FloatType newZ)
+void quaternion::set(double newW, double newX, double newY, double newZ)
 {
     w = newW, x = newX, y = newY, z = newZ;
 }
 
 void quaternion::set(const matrix33& m33)
 {
-    static const FloatType eps = std::numeric_limits<FloatType>::epsilon();
-    static const FloatType quater = static_cast<FloatType>(0.25);
+    static const double eps = std::numeric_limits<double>::epsilon();
+    static const double quater = static_cast<double>(0.25);
 
-    FloatType onePlusTrace = m33.trace() + (FloatType)1.0;
+    double onePlusTrace = m33.trace() + 1.0;
 
     if (onePlusTrace > eps) 
     {
-        FloatType S = std::sqrt(onePlusTrace) * (FloatType)2.0;
+        double S = std::sqrt(onePlusTrace) * 2.0;
         w = quater * S;
         x = (m33(2, 1) - m33(1, 2)) / S;
         y = (m33(0, 2) - m33(2, 0)) / S;
@@ -40,7 +40,7 @@ void quaternion::set(const matrix33& m33)
     }
     else if (m33(0, 0) > m33(1, 1) && m33(0, 0) > m33(2, 2))
     {
-        FloatType S = std::sqrt((FloatType)1.0 + m33(0, 0) - m33(1, 1) - m33(2, 2)) * (FloatType)2.0;
+        double S = std::sqrt(1.0 + m33(0, 0) - m33(1, 1) - m33(2, 2)) * 2.0;
         w = (m33(2, 1) - m33(1, 2)) / S;
         x = quater * S;
         y = (m33(0, 1) + m33(1, 0)) / S;
@@ -48,7 +48,7 @@ void quaternion::set(const matrix33& m33)
     }
     else if (m33(1, 1) > m33(2, 2)) 
     {
-        FloatType S = std::sqrt((FloatType)1.0 + m33(1, 1) - m33(0, 0) - m33(2, 2)) * (FloatType)2.0;
+        double S = std::sqrt(1.0 + m33(1, 1) - m33(0, 0) - m33(2, 2)) * 2.0;
         w = (m33(0, 2) - m33(2, 0)) / S;
         x = (m33(0, 1) + m33(1, 0)) / S;
         y = quater * S;
@@ -56,7 +56,7 @@ void quaternion::set(const matrix33& m33)
     }
     else 
     {
-        FloatType S = std::sqrt((FloatType)1.0 + m33(2, 2) - m33(0, 0) - m33(1, 1)) * (FloatType)2.0;
+        double S = std::sqrt(1.0 + m33(2, 2) - m33(0, 0) - m33(1, 1)) * 2.0;
         w = (m33(1, 0) - m33(0, 1)) / S;
         x = (m33(0, 2) + m33(2, 0)) / S;
         y = (m33(1, 2) + m33(2, 1)) / S;
@@ -71,20 +71,20 @@ void quaternion::setIdentity()
 
 matrix33 quaternion::matrix3() const
 {
-    FloatType _2xx = (FloatType)2.0 * x * x;
-    FloatType _2yy = (FloatType)2.0 * y * y;
-    FloatType _2zz = (FloatType)2.0 * z * z;
-    FloatType _2xy = (FloatType)2.0 * x * y;
-    FloatType _2xz = (FloatType)2.0 * x * z;
-    FloatType _2xw = (FloatType)2.0 * x * w;
-    FloatType _2yz = (FloatType)2.0 * y * z;
-    FloatType _2yw = (FloatType)2.0 * y * w;
-    FloatType _2zw = (FloatType)2.0 * z * w;
+    double _2xx = 2.0 * x * x;
+    double _2yy = 2.0 * y * y;
+    double _2zz = 2.0 * z * z;
+    double _2xy = 2.0 * x * y;
+    double _2xz = 2.0 * x * z;
+    double _2xw = 2.0 * x * w;
+    double _2yz = 2.0 * y * z;
+    double _2yw = 2.0 * y * w;
+    double _2zw = 2.0 * z * w;
 
     matrix33 m(
-        (FloatType)1.0 - _2yy - _2zz, _2xy - _2zw, _2xz + _2yw,
-        _2xy + _2zw, (FloatType)1.0 - _2zz - _2xx, _2yz - _2xw,
-        _2xz - _2yw, _2yz + _2xw, (FloatType)1.0 - _2yy - _2xx);
+        1.0 - _2yy - _2zz, _2xy - _2zw, _2xz + _2yw,
+        _2xy + _2zw, 1.0 - _2zz - _2xx, _2yz - _2xw,
+        _2xz - _2yw, _2yz + _2xw, 1.0 - _2yy - _2xx);
 
     return m;
 }
@@ -92,7 +92,7 @@ matrix33 quaternion::matrix3() const
 
 quaternion quaternion::inverse() const
 {
-    FloatType denom = w * w + x * x + y * y + z * z;
+    double denom = w * w + x * x + y * y + z * z;
     return quaternion(w / denom, -x / denom, -y / denom, -z / denom);
 }
