+++ fluidSimLib/vertexCenteredScalarGrid3.h
@@ -10,7 +10,7 @@ public:
     vertexCenteredScalarGrid3();
 
     vertexCenteredScalarGrid3(const size3& resolution, const vector3& gridSpacing = vector3(1.0, 1.0, 1.0),
-        const vector3& origin = vector3(), FloatType initValue = 0.0);
+        const vector3& origin = vector3(), double initValue = 0.0);
 
     virtual ~vertexCenteredScalarGrid3();
